update decimal to bigInteger

+++ database/migrations/2024_04_01_031731_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->text('description');
             $table->unsignedBigInteger('category_id')->nullable();
             $table->foreign('category_id')->references('id')->on('categories');
-            $table->decimal('price');
+            $table->bigInteger('price');
             $table->integer('Qty');            
             $table->timestamps();
         });
